requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_llvm
absl-py=0.9.0=py36_0
appdirs=1.4.3=py_1
asciitree=0.3.3=py_2
asn1crypto=1.3.0=py36_0
astroid=2.3.3=py36_0
attrs=19.3.0=py_0
backcall=0.1.0=py_0
basemap=1.2.2=py36h90df80e_3
basemap-data-hires=1.2.0=0
bayesian-optimization=1.1.0=py_0
bcrypt=3.1.7=py36h8c4c3a4_1
blas=1.0=mkl
bleach=3.1.0=py_0
blessings=1.7=py36h9f0ad1d_1003
bokeh=2.0.1=py36h9f0ad1d_0
bottleneck=1.3.2=py36heb32a55_1
bzip2=1.0.8=h7b6447c_0
c-ares=1.15.0=h516909a_1001
ca-certificates=2020.12.5=ha878542_0
captum=0.3.0=0
cartopy=0.18.0=py36hb73674a_8
cdo=1.9.9=h04f682f_14
cdsapi=0.1.6=py_0
certifi=2020.12.5=py36h5fab9bb_1
cffi=1.13.2=py36h8022711_0
cftime=1.0.4.2=py36hc1659b7_0
chardet=3.0.4=py36_1003
click=7.0=py_0
cloog=0.18.0=0
cloudpickle=1.3.0=py_0
cryptography=2.8=py36h72c5cf5_1
cudatoolkit=10.2.89=h8f6ccaa_8
curl=7.71.1=he644dc0_3
cycler=0.10.0=py_2
cython=0.29.17=py36he6710b0_0
cytoolz=0.10.1=py36h7b6447c_0
dask=2.15.0=py_0
dask-core=2.15.0=py_0
dataclasses=0.8=pyh787bdff_0
dbus=1.13.12=h746ee38_0
decorator=4.4.1=py_0
defusedxml=0.6.0=py_0
distributed=2.15.2=py36h9f0ad1d_0
eccodes=2.19.1=hea64003_0
entrypoints=0.3=py36_1000
esmf=8.0.0=mpi_mpich_h213fab7_107
esmpy=8.0.0=mpi_mpich_py36ha9b28fa_101
expat=2.2.6=he6710b0_0
fasteners=0.14.1=py_3
ffmpeg=4.2=h167e202_0
fftw=3.3.8=nompi_h7f3a6c3_1110
findspark=1.3.0=py_1
fontconfig=2.13.1=h86ecdb6_1001
freetype=2.9.1=h8a8886c_1
fsspec=0.6.2=py_0
future=0.18.2=py36h5fab9bb_3
geos=3.8.1=he1b5a44_0
glib=2.63.1=h5a9c865_0
gmp=6.2.0=he1b5a44_0
gnutls=3.6.5=hd3a4fd2_1002
gpustat=0.6.0=pyhd8ed1ab_1
grpcio=1.23.0=py36h769ab6c_1
gst-plugins-base=1.14.5=h0935bb2_0
gstreamer=1.14.5=h36ae1b5_0
hdf4=4.2.13=h3ca952b_2
hdf5=1.10.6=mpi_mpich_hc41e3f9_1011
heapdict=1.0.1=py_0
hyperopt=0.2.5=pyh9f0ad1d_0
icu=64.2=he1b5a44_1
idna=2.8=py36_1000
importlib_metadata=1.4.0=py36_0
intel-openmp=2019.4=243
ipykernel=5.1.3=py36h5ca1d4c_0
ipympl=0.4.1=py_0
ipython=7.11.1=py36h5ca1d4c_0
ipython_genutils=0.2.0=py_1
ipytree=0.2.1=pyhd8ed1ab_0
ipywidgets=7.6.3=pyhd3eb1b0_1
isl=0.12.2=0
isort=4.3.21=py36_0
jasper=1.900.1=h07fcdf6_1006
jedi=0.16.0=py36_0
jinja2=2.10.3=py_0
joblib=0.14.1=py_0
jpeg=9d=h36c2ea0_0
json5=0.8.5=py_0
jsonschema=3.2.0=py36_0
jupyter_client=5.3.4=py36_1
jupyter_core=4.6.1=py36_0
jupyterlab=1.2.5=py_0
jupyterlab_server=1.0.6=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.1.0=py36hc9558a2_0
krb5=1.17.2=h926e7f8_0
lame=3.100=h14c3975_1001
lazy-object-proxy=1.4.3=py36h7b6447c_0
ld_impl_linux-64=2.33.1=h53a641e_7
libaec=1.0.4=hf484d3e_0
libblas=3.9.0=8_mkl
libcblas=3.9.0=8_mkl
libcurl=7.71.1=hcdd3856_3
libedit=3.1.20191231=he28a2e2_2
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.3.0=h5dbcf3e_17
libgfortran-ng=7.5.0=h14aa051_18
libgfortran4=7.5.0=h14aa051_18
libgomp=9.3.0=h5dbcf3e_17
libiconv=1.15=h516909a_1005
liblapack=3.9.0=8_mkl
liblapacke=3.9.0=8_mkl
libnetcdf=4.7.4=mpi_mpich_hfd9c5b6_4
libopenblas=0.3.7=h5ec1e0e_6
libpng=1.6.37=hbc83047_0
libprotobuf=3.14.0=h780b84a_0
libsodium=1.0.17=h516909a_0
libssh2=1.9.0=hab1572f_5
libstdcxx-ng=9.3.0=h6de172a_18
libtiff=4.1.0=hc3755c2_3
libuuid=2.32.1=h14c3975_1000
libuv=1.41.0=h7f98852_0
libxcb=1.13=h1bed415_1
libxml2=2.9.10=hee79883_0
llvm-openmp=11.0.1=h4bd325d_0
locket=0.2.0=py36_1
lz4-c=1.8.3=he1b5a44_1001
markdown=3.2.1=py_0
markupsafe=1.1.1=py36h516909a_0
matplotlib=3.1.2=py36_1
matplotlib-base=3.1.2=py36h250f245_1
mccabe=0.6.1=py36_1
metpy=0.12.1=py_0
mistune=0.8.4=py36h516909a_1000
mkl=2020.4=h726a3e6_304
mkl-service=2.3.0=py36h8c4c3a4_2
mkl_fft=1.3.0=py36h92226af_1
mkl_random=1.2.0=py36h7c3b610_1
monotonic=1.5=py_0
more-itertools=8.1.0=py_0
mpc=1.1.0=4
mpfr=3.1.5=0
mpi=1.0=mpich
mpi4py=3.0.3=py36h094ef99_1
mpich=3.3.2=hc856adb_0
msgpack-python=0.6.1=py36hfd86e86_1
nbconvert=5.6.1=py36_0
nbformat=5.0.4=py_0
nc-time-axis=1.2.0=py_1
ncurses=6.2=h58526e2_4
netcdf-fortran=4.5.3=mpi_mpich_h3923e1a_1
netcdf4=1.5.5.1=nompi_py36hc29086f_101
nettle=3.4.1=h1bed415_1002
networkx=2.4=py_0
ninja=1.9.0=py36hfd86e86_0
nodejs=10.13.0=he6710b0_0
notebook=6.0.3=py36_0
numcodecs=0.7.2=py36h831f99a_1
numpy=1.18.1=py36h95a1406_0
numpy-base=1.18.5=py36hde5b4d6_0
nvidia-ml=7.352.0=py_0
olefile=0.46=py_0
onnx=1.8.0=py36h6889425_1
openh264=1.8.0=hdbcaa40_1000
openjdk=8.0.152=h7b6447c_3
openssl=1.1.1j=h7f98852_0
ossuuid=1.6.2=hf484d3e_1000
owslib=0.19.0=py_2
packaging=20.1=py_0
pandas=1.0.3=py36h0573a6f_0
pandoc=2.9.1.1=0
pandocfilters=1.4.2=py_1
paramiko=2.7.1=py36_0
parso=0.5.2=py_0
partd=1.1.0=py_0
patsy=0.5.1=py36_0
pcre=8.43=he6710b0_0
pexpect=4.8.0=py36_0
pickleshare=0.7.5=py36_1000
pillow=7.0.0=py36hefe7db6_0
pint=0.10.1=py_0
pip=20.3.3=py36h06a4308_0
pooch=1.0.0=py_0
proj=7.2.0=h8b9fe22_0
prometheus_client=0.7.1=py_0
prompt_toolkit=3.0.2=py_0
protobuf=3.14.0=py36hc4f0c31_1
psutil=5.6.7=py36h7b6447c_0
ptyprocess=0.6.0=py_1001
py4j=0.10.7=py_1
pycparser=2.19=py_0
pyepsg=0.4.0=py_0
pygments=2.5.2=py_0
pykdtree=1.3.1=py36hc1659b7_1002
pylint=2.4.4=py36_0
pymongo=3.11.0=py36ha357f81_4
pynacl=1.3.0=py36h516909a_1001
pyopenssl=19.1.0=py36_0
pyparsing=2.4.6=py_0
pyproj=1.9.5.1=py36_0
pyqt=5.9.2=py36hcca6a23_4
pyrsistent=0.15.7=py36h516909a_0
pysftp=0.2.9=py_1
pyshp=2.1.0=py_0
pysocks=1.7.1=py36_0
pyspark=2.4.6=py_0
python=3.6.7=h381d211_1004
python-cdo=1.5.3=py_0
python-dateutil=2.8.1=py_0
python_abi=3.6=1_cp36m
pytorch=1.7.0=py3.6_cuda10.2.89_cudnn7.6.5_0
pytz=2019.3=py_0
pyyaml=5.3=py36h7b6447c_0
pyzmq=18.1.1=py36h1768529_0
qt=5.9.7=h0c104cb_3
readline=7.0=h7b6447c_5
requests=2.22.0=py36_1
scikit-learn=0.22.1=py36hcdab131_1
scipy=1.5.2=py36h976291a_2
seaborn=0.9.0=pyh91ea838_1
send2trash=1.5.0=py_0
setuptools=45.1.0=py36_0
shapely=1.7.1=py36h558fe5e_1
sip=4.19.8=py36hf484d3e_1000
six=1.14.0=py36_0
sortedcontainers=2.1.0=py36_0
sqlite=3.33.0=h62c20be_0
statsmodels=0.10.1=py36hdd07704_0
tblib=1.6.0=py_0
tensorboard=1.15.0=py36_0
terminado=0.8.3=py36_0
testpath=0.4.4=py_0
tk=8.6.10=hbc83047_0
toolz=0.10.0=py_0
torchvision=0.8.1=py36_cu102
tornado=6.0.3=py36h516909a_0
tqdm=4.54.1=pyhd8ed1ab_0
traitlets=4.3.3=py36_0
typed-ast=1.4.1=py36h7b6447c_0
typing-extensions=3.7.4.3=0
typing_extensions=3.7.4.3=py_0
udunits2=2.2.27.6=h4e0c4b3_1001
urllib3=1.25.7=py36_0
wcwidth=0.1.8=py_0
webencodings=0.5.1=py_1
werkzeug=1.0.0=py_0
wget=1.20.1=h20c2e04_0
wheel=0.33.6=py36_0
widgetsnbextension=3.5.1=py36_0
wrapt=1.11.2=py36h7b6447c_0
wrf-python=1.3.2=py36h7eb8c7e_1
x264=1!152.20180806=h14c3975_0
xarray=0.15.1=py_0
xesmf=0.3.0=py_0
xz=5.2.4=h14c3975_4
yaml=0.1.7=had09818_2
zarr=2.5.0=py_0
zeromq=4.3.2=he1b5a44_2
zict=1.0.0=py_0
zipp=2.0.0=py_2
zlib=1.2.11=h7b6447c_3
zstd=1.4.4=h3b9ef0a_1
